# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Self-hostable multi-device AI assistant API",
        "title": "openCrow API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/conversations/{id}/messages/{msgId}/regenerate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "conversations"
                ],
                "summary": "Regenerate an assistant message via SSE stream",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "msgId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE stream of delta/tool_call/tool_result/done/error events",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/orchestrator/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator"
                ],
                "summary": "Run a full orchestrator completion (blocking)",
                "parameters": [
                    {
                        "description": "Message and conversation ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CompleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CompleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/orchestrator/stream": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "orchestrator"
                ],
                "summary": "Stream an orchestrator completion via SSE",
                "parameters": [
                    {
                        "description": "Message and conversation ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CompleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE stream of delta/tool_call/tool_result/done/error events",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/realtime/last": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator"
                ],
                "summary": "Get the last realtime event for the current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.CompleteRequest": {
            "type": "object",
            "properties": {
                "conversationId": {
                    "type": "string"
                },
                "maxRetries": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "providerOrder": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "api.CompleteResponse": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "integer"
                },
                "output": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "trace": {
                    "$ref": "#/definitions/api.CompletionTraceResponse"
                },
                "usage": {
                    "$ref": "#/definitions/api.TokenUsageDTO"
                }
            }
        },
        "api.CompletionTraceResponse": {
            "type": "object",
            "properties": {
                "providerAttempts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ProviderAttemptDTO"
                    }
                },
                "runtimeActions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.RuntimeActionDTO"
                    }
                },
                "toolCalls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ToolCallDTO"
                    }
                }
            }
        },
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.ProviderAttemptDTO": {
            "type": "object",
            "properties": {
                "attempt": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.RuntimeActionDTO": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "api.TokenUsageDTO": {
            "type": "object",
            "properties": {
                "completionTokens": {
                    "type": "integer"
                },
                "promptTokens": {
                    "type": "integer"
                },
                "totalTokens": {
                    "type": "integer"
                }
            }
        },
        "api.ToolCallDTO": {
            "type": "object",
            "properties": {
                "arguments": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "name": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}